Drop redundant searchsorted/unravel_index int32 casts
`jnp.searchsorted` already returns int32 even with `jax_enable_x64`, so
the four `.astype(jnp.int32)` casts in `grids/coordinates.py`,
`grids/piecewise.py` (×2), and `simulation/simulate.py:_compute_starting_periods`
were no-ops at the dtype level — but they sat between an integer-producing
op and its index-consumer inside vmap'd interpolation kernels, breaking
XLA's fusion and forcing the intermediate to materialise as a top-level
GPU buffer per (period, regime, state). Likewise, the `unravel_index`
output in `_lookup_values_from_indices` is consumed immediately by
`grid[index]`, which accepts int64 fine — the cast served no purpose.

Keeps the argmax cast on the solve path (real int64→int32 narrowing),
the boundary casts at error/validation paths, and the AOT-relevant
casts in `pandas_utils` and the `subject_regime_ids` sentinel.

+++ src/lcm/grids/coordinates.py
@@ -188,7 +188,7 @@ def get_irreg_coordinate(
     n_points = len(points)
 
     # Find the index of the first point greater than value
-    idx_upper = jnp.searchsorted(points, value, side="right").astype(jnp.int32)
+    idx_upper = jnp.searchsorted(points, value, side="right")
 
     # Clamp to valid range for interpolation
     idx_upper = jnp.clip(idx_upper, 1, n_points - 1)
